#3 refactor: seperate DesktopHeader and MobilHeader
