+++ src/components/header/Header.tsx
@@ -1,0 +1,9 @@
+import useWindowWidth from "../../hooks/useWindowWidth";
+import DesktopHeader from "./DesktopHeader";
+import MobileHeader from "./MobileHeader";
+
+export default function Header() {
+  const windowWidth = useWindowWidth();
+
+  return <>{windowWidth >= 730 ? <DesktopHeader /> : <MobileHeader />}</>;
+}
