+++ src/components/layout/ProjectLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router";
-import Header from "../Header";
+import Header from "../header/Header";
 
 export default function ProjectLayout() {
   return (
